requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
pytorch-lightning==1.5.10
hydra-core
omegaconf
wandb
einops
opt_einsum
cmake # For pykeops support
pykeops # If there are installation problems with pykeops==2.x, try pykeops==1.5
transformers # For some schedulers
rotary-embedding-torch # For rotary position embeddings

# Model specific packges
# pytorch-fast-transformers # for Performer

# Dataset specific packages
torchtext # LRA
torchdata
torchmetrics

datasets # LRA
# gluonts
# timm
# lightning-bolts
sktime # BIDMC
